VisualOdometry/requirements.txt
numpy==2.2.2
opencv-python==4.11.0.86
PyYAML==6.0.2
